+++ src/ragger/navigator/stax_navigator.py
@@ -62,6 +62,7 @@ class StaxNavigator(Navigator):
             NavInsID.USE_CASE_CHOICE_CONFIRM: screen.choice.confirm,
             NavInsID.USE_CASE_CHOICE_REJECT: screen.choice.reject,
             NavInsID.USE_CASE_STATUS_WAIT: screen.status.wait,
+            NavInsID.USE_CASE_STATUS_DISMISS: screen.status.dismiss,
             NavInsID.USE_CASE_REVIEW_TAP: screen.review.tap,
             NavInsID.USE_CASE_REVIEW_PREVIOUS: screen.review.previous,
             NavInsID.USE_CASE_REVIEW_REJECT: screen.review.reject,